[refactor] AnswerService 디렉토리 생성
