+++ src/main/java/qp/official/qp/service/AnswerService/AnswerQueryServiceImpl.java
@@ -1,6 +1,5 @@
-package qp.official.qp.service;
+package qp.official.qp.service.AnswerService;
 
-import java.util.List;
 import java.util.NoSuchElementException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
