+++ src/main/java/qp/official/qp/web/controller/AnswerRestController.java
@@ -10,8 +10,8 @@ import qp.official.qp.apiPayload.ApiResponse;
 import qp.official.qp.apiPayload.code.status.SuccessStatus;
 import qp.official.qp.converter.AnswerConverter;
 import qp.official.qp.domain.Answer;
-import qp.official.qp.service.AnswerCommandService;
-import qp.official.qp.service.AnswerQueryService;
+import qp.official.qp.service.AnswerService.AnswerCommandService;
+import qp.official.qp.service.AnswerService.AnswerQueryService;
 import qp.official.qp.validation.annotation.ExistAnswer;
 import qp.official.qp.validation.annotation.ExistQuestion;
 import qp.official.qp.web.dto.AnswerRequestDTO;
